Update to support Phoenix 1.7
